requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.36
python-telegram-bot==21.6
python-dotenv==1.0.1
openai>=1.30.0
requests>=2.31.0
